+++ amplipi/app.py
@@ -457,6 +457,13 @@ def announce(announcement: models.Announcement, ctrl: Api = Depends(get_ctrl)) -
   """ Make an announcement """
   return code_response(ctrl, ctrl.announce(announcement))
 
+# Info
+
+@api.get('/api/info', tags=['status'])
+def get_info(ctrl: Api = Depends(get_ctrl)) -> models.Info:
+  """ Get additional information """
+  return code_response(ctrl, ctrl.get_info())
+
 # include all routes above
 
 app.include_router(api)