+++ amplipi/ctrl.py
@@ -38,6 +38,8 @@ import amplipi.utils as utils
 
 _DEBUG_API = False # print out a graphical state of the api after each call
 
+USER_CACHE_DIR = os.path.join(os.path.expanduser('~'), '.cache', 'amplipi')
+
 @wrapt.decorator
 def save_on_success(wrapped, instance: 'Api', args, kwargs):
   """ Check if a ctrl API call is successful and saves the state if so """
@@ -194,14 +196,21 @@ class Api:
       self.status = models.Status.parse_obj(self.DEFAULT_CONFIG)
       self.save()
 
+    # TODO: detect missing sources
+
+    # populate system info
+    self._online_cache = utils.TimeBasedCache(self._check_is_online, 5, 'online')
+    self._latest_release_cache = utils.TimeBasedCache(self._check_latest_release, 3600, 'latest release')
     self.status.info = models.Info(
       mock_ctrl=self._mock_hw,
       mock_streams=self._mock_streams,
       config_file=self.config_file,
-      version=utils.detect_version()
+      version=utils.detect_version(),
     )
-
-    # TODO: detect missing sources
+    for major, minor, ghash, dirty in self._rt.read_versions():
+      fw_info = models.FirmwareInfo(version=f'{major}.{minor}', git_hash=f'{ghash:x}', git_dirty=dirty)
+      self.status.info.fw.append(fw_info)
+    self._update_sys_info() # TODO: does sys info need to be updated at init time?
 
     # detect missing zones
     if self._mock_hw:
@@ -319,6 +328,31 @@ class Api:
       inputs['stream={}'.format(stream.id)] = f'{stream.name} - {stream.type}'
     return inputs
 
+  def _check_is_online(self) -> bool:
+    online = False
+    try:
+      with open(os.path.join(USER_CACHE_DIR, 'online'), encoding='utf-8') as fonline:
+        online = 'online' in fonline.readline()
+    except Exception as exc:
+      pass
+    return online
+
+  def _check_latest_release(self) -> str:
+    release = 'unknown'
+    try:
+      with open(os.path.join(USER_CACHE_DIR, 'latest_release'), encoding='utf-8') as flatest:
+        release = flatest.readline().strip()
+    except Exception as exc:
+      pass
+    return release
+
+  def _update_sys_info(self, throttled = True) -> None:
+    """Update current system information"""
+    if self.status.info is None:
+      raise Exception("No info generated, system in a bad state")
+    self.status.info.online = self._online_cache.get(throttled)
+    self.status.info.latest_release = self._latest_release_cache.get(throttled)
+
   def get_state(self) -> models.Status:
     """ get the system state """
     # update the state with the latest stream info
@@ -335,11 +369,19 @@ class Api:
           stream.__dict__[field] = stream_inst.__dict__[field]
       streams.append(stream)
     self.status.streams = streams
+    self._update_sys_info()
     # update source's info
     # TODO: stream/source info should be updated in a background thread
     for src in self.status.sources:
       self._update_src_info(src)
     return self.status
+
+  def get_info(self) -> models.Info:
+    """ Get the system information """
+    self._update_sys_info()
+    if self.status.info is None:
+      raise Exception("No info generated, system in a bad state")
+    return self.status.info
 
   def get_items(self, tag: str) -> Optional[List[models.Base]]:
     """ Gets one of the lists of elements contained in status named by @t (or t's plural